+++ adapters/InMobi/InMobiAdapter/GADMAdapterInMobiUtils.m
@@ -1,7 +1,16 @@
+// Copyright 2019 Google LLC
 //
-//  GADMAdapterInMobiUtils.m
-//  Adapter
-//  Copyright © 2019 Google. All rights reserved.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 //
 
 #import "GADMAdapterInMobiUtils.h"
@@ -16,6 +25,8 @@
 
 /// Sets up additional InMobi targeting information from the specified |extras|.
 void GADMAdapterInMobiSetTargetingFromExtras(GADInMobiExtras *_Nullable extras);
+
+void GADMAdapterInMobiSetIsAgeRestricted(NSNumber *_Nullable isRestricted);
 
 /// Sets additional InMobi parameters to |requestParameters| from the specified |extras|.
 NSDictionary<NSString *, id> *_Nonnull GADMAdapterInMobiAdditonalParametersFromInMobiExtras(
@@ -82,7 +93,7 @@ NSError *_Nullable GADMAdapterInMobiValidatePlacementIdentifier(
 
   return GADMAdapterInMobiErrorWithCodeAndDescription(
       GADMAdapterInMobiErrorInvalidServerParameters,
-      @"[InMobi] Error - Placement ID not specified.");
+      @"GADMediationAdapterInMobi - Error : Placement ID not specified.");
 }
 
 void GADMAdapterInMobiSetTargetingFromExtras(GADInMobiExtras *_Nullable extras) {
@@ -113,6 +124,15 @@ void GADMAdapterInMobiSetTargetingFromExtras(GADInMobiExtras *_Nullable extras) 
   }
 }
 
+void GADMAdapterInMobiLog(NSString *_Nonnull format, ...) {
+  va_list arguments;
+  va_start(arguments, format);
+  NSString *log = [[NSString alloc] initWithFormat:format arguments:arguments];
+  va_end(arguments);
+
+  NSLog(@"GADMediationAdapterInMobi - %@", log);
+}
+
 void GADMAdapterInMobiSetTargetingFromConnector(id<GADMAdNetworkConnector> _Nonnull connector) {
   if (connector.userGender == kGADGenderMale) {
     [IMSdk setGender:kIMSDKGenderMale];
@@ -128,11 +148,19 @@ void GADMAdapterInMobiSetTargetingFromConnector(id<GADMAdNetworkConnector> _Nonn
   }
 
   GADMAdapterInMobiSetTargetingFromExtras([connector networkExtras]);
+  GADMAdapterInMobiSetIsAgeRestricted(connector.childDirectedTreatment);
 }
 
 void GADMAdapterInMobiSetTargetingFromAdConfiguration(
     GADMediationAdConfiguration *_Nonnull adConfig) {
   GADMAdapterInMobiSetTargetingFromExtras(adConfig.extras);
+  GADMAdapterInMobiSetIsAgeRestricted(adConfig.childDirectedTreatment);
+}
+
+void GADMAdapterInMobiSetIsAgeRestricted(NSNumber *_Nullable isRestricted) {
+  if ([isRestricted isEqual:@1]) {
+    [IMSdk setIsAgeRestricted:true];
+  }
 }
 
 NSDictionary<NSString *, id> *_Nonnull GADMAdapterInMobiAdditonalParametersFromInMobiExtras(