+++ adapters/Vungle/VungleAdapter/Bidding/GADMediationVungleInterstitial.m
@@ -72,7 +72,9 @@
 - (void)requestInterstitialAd {
   if (!self.desiredPlacement.length) {
     NSError *error = GADMAdapterVungleErrorWithCodeAndDescription(
-        GADMAdapterVungleErrorInvalidServerParameters, @"Placement ID not specified.");
+        GADMAdapterVungleErrorInvalidServerParameters,
+        @"Missing or invalid Placement ID configured for this ad source instance in the AdMob or "
+        @"Ad Manager UI.");
     _adLoadCompletionHandler(nil, error);
     return;
   }
@@ -160,7 +162,7 @@
 }
 
 - (void)didViewAd {
-  // Do nothing.
+  [_delegate reportImpression];
 }
 
 - (void)willCloseAd {