+++ adapters/Tapjoy/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Tapjoy iOS Mediation Adapter Changelog
 
+#### Version 12.11.1.0 (In progress)
+- Verified compatibility with Tapjoy SDK 12.11.1.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.13.0.
+- Tapjoy SDK version 12.11.1.
+
 #### [Version 12.11.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/tapjoy/TapjoyAdapter-12.11.0.0.zip)
 - Updated the adapter to use the `didRewardUser` API.
 - Now requires Google Mobile Ads SDK version 9.8.0 or higher.