+++ adapters/Vungle/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Vungle iOS Mediation Adapter Changelog
 
+####  [Version 6.12.1.1](https://dl.google.com/googleadmobadssdk/mediation/ios/vungle/VungleAdapter-6.12.1.1.zip)
+- Verified compatibility with Vungle SDK 6.12.1.
+- Added bidding support for banner ad format.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.14.0.
+- Vungle SDK version 6.12.1.
+
 #### [Version 6.12.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/vungle/VungleAdapter-6.12.1.0.zip)
 - Verified compatibility with Vungle SDK 6.12.1.